requirements.txt
annotated-types==0.6.0
anyio==3.7.1
exceptiongroup==1.2.0
fastapi==0.105.0
future==0.18.3
idna==3.6
libtorrent==2.0.9
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.1.0
python-ffmpeg==2.0.9
pytube==15.0.0
sniffio==1.3.0
starlette==0.27.0
torrentp==0.1.6
typing_extensions==4.9.0
